+++ packages/kit/src/runtime/client/start.js
@@ -1,31 +1,20 @@
 import { DEV } from 'esm-env';
 import { create_client } from './client.js';
 import { init } from './singletons.js';
-import { set_assets, set_version, set_public_env } from '../shared.js';
 
 /**
- * @param {{
- *   assets: string;
- *   env: Record<string, string>;
- *   hydrate: Parameters<import('./types').Client['_hydrate']>[0];
- *   target: HTMLElement;
- *   version: string;
- * }} opts
+ * @param {import('./types').SvelteKitApp} app
+ * @param {HTMLElement} target
+ * @param {Parameters<import('./types').Client['_hydrate']>[0]} [hydrate]
  */
-export async function start({ assets, env, hydrate, target, version }) {
-	set_public_env(env);
-	set_assets(assets);
-	set_version(version);
-
+export async function start(app, target, hydrate) {
 	if (DEV && target === document.body) {
 		console.warn(
 			`Placing %sveltekit.body% directly inside <body> is not recommended, as your app may break for users who have certain browser extensions installed.\n\nConsider wrapping it in an element:\n\n<div style="display: contents">\n  %sveltekit.body%\n</div>`
 		);
 	}
 
-	const client = create_client({
-		target
-	});
+	const client = create_client(app, target);
 
 	init({ client });