+++ packages/kit/src/runtime/env/dynamic/public.js
@@ -1,1 +1,1 @@
-export { public_env as env } from '../../shared.js';
+export { public_env as env } from '../../shared-server.js';