+++ packages/kit/src/runtime/server/utils.js
@@ -1,9 +1,8 @@
-import * as devalue from 'devalue';
 import { json, text } from '../../exports/index.js';
 import { coalesce_to_error } from '../../utils/error.js';
 import { negotiate } from '../../utils/http.js';
 import { HttpError } from '../control.js';
-import { fix_stack_trace } from '../shared.js';
+import { fix_stack_trace } from '../shared-server.js';
 
 /** @param {any} body */
 export function is_pojo(body) {
@@ -16,11 +15,6 @@ export function is_pojo(body) {
 
 	return true;
 }
-
-/** @type {import('types').SSRErrorPage} */
-export const GENERIC_ERROR = {
-	id: '__error'
-};
 
 /**
  * @param {Partial<Record<import('types').HttpMethod, any>>} mod
@@ -41,7 +35,7 @@ export function method_not_allowed(mod, method) {
 export function allowed_methods(mod) {
 	const allowed = [];
 
-	for (const method in ['GET', 'POST', 'PUT', 'PATCH', 'DELETE']) {
+	for (const method in ['GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'OPTIONS']) {
 		if (method in mod) allowed.push(method);
 	}
 
@@ -99,7 +93,7 @@ export async function handle_error_and_jsonify(event, options, error) {
 	if (error instanceof HttpError) {
 		return error.body;
 	} else {
-		if (__SVELTEKIT_DEV__) {
+		if (__SVELTEKIT_DEV__ && typeof error == 'object') {
 			error = new Proxy(error, {
 				get: (target, property) => {
 					if (property === 'stack') {
@@ -148,31 +142,23 @@ export function clarify_devalue_error(event, error) {
 	return error.message;
 }
 
-/** @param {import('types').ServerDataNode | import('types').ServerDataSkippedNode | import('types').ServerErrorNode | null} node */
-export function serialize_data_node(node) {
-	if (!node) return 'null';
-
-	if (node.type === 'error' || node.type === 'skip') {
-		return JSON.stringify(node);
-	}
-
-	const stringified = devalue.stringify(node.data);
-
+/**
+ * @param {import('types').ServerDataNode} node
+ */
+export function stringify_uses(node) {
 	const uses = [];
 
-	if (node.uses.dependencies.size > 0) {
+	if (node.uses && node.uses.dependencies.size > 0) {
 		uses.push(`"dependencies":${JSON.stringify(Array.from(node.uses.dependencies))}`);
 	}
 
-	if (node.uses.params.size > 0) {
+	if (node.uses && node.uses.params.size > 0) {
 		uses.push(`"params":${JSON.stringify(Array.from(node.uses.params))}`);
 	}
 
-	if (node.uses.parent) uses.push(`"parent":1`);
-	if (node.uses.route) uses.push(`"route":1`);
-	if (node.uses.url) uses.push(`"url":1`);
+	if (node.uses?.parent) uses.push(`"parent":1`);
+	if (node.uses?.route) uses.push(`"route":1`);
+	if (node.uses?.url) uses.push(`"url":1`);
 
-	return `{"type":"data","data":${stringified},"uses":{${uses.join(',')}}${
-		node.slash ? `,"slash":${JSON.stringify(node.slash)}` : ''
-	}}`;
+	return `"uses":{${uses.join(',')}}`;
 }