+++ packages/kit/test/apps/basics/src/hooks.client.js
@@ -1,3 +1,7 @@
+import { env } from '$env/dynamic/public';
+
+window.PUBLIC_DYNAMIC = env.PUBLIC_DYNAMIC;
+
 /** @type{import("@sveltejs/kit").HandleClientError} */
 export function handleError({ error, event }) {
 	return event.url.pathname.endsWith('404-fallback')