+++ packages/kit/test/apps/basics/test/client.test.js
@@ -544,7 +544,7 @@ test.describe('data-sveltekit attributes', () => {
 
 		const module = process.env.DEV
 			? `${baseURL}/src/routes/data-sveltekit/preload-data/target/+page.svelte`
-			: `${baseURL}/_app/immutable/components/pages/data-sveltekit/preload-data/target/_page`;
+			: `${baseURL}/_app/immutable/entry/data-sveltekit-preload-data-target-page`;
 
 		await page.goto('/data-sveltekit/preload-data');
 		await page.locator('#one').dispatchEvent('mousemove');
@@ -636,10 +636,17 @@ test.describe('Content negotiation', () => {
 	});
 });
 
-test.describe('env in app.html', () => {
-	test('can access public env', async ({ page }) => {
+test.describe('env', () => {
+	test('can access public env in app.html', async ({ page }) => {
 		await page.goto('/');
 		expect(await page.locator('body').getAttribute('class')).toContain('groovy');
+	});
+
+	test('can access public env in hooks.client.js', async ({ page }) => {
+		await page.goto('/');
+		expect(await page.evaluate(() => window.PUBLIC_DYNAMIC)).toBe(
+			'accessible anywhere/evaluated at run time'
+		);
 	});
 });
 
@@ -670,4 +677,77 @@ test.describe('Snapshots', () => {
 		await page.reload();
 		expect(await page.locator('input').inputValue()).toBe('works for reloads');
 	});
+});
+
+test.describe('Streaming', () => {
+	test('Works for universal load functions (client nav)', async ({ page }) => {
+		await page.goto('/streaming');
+		page.click('[href="/streaming/universal"]');
+
+		await expect(page.locator('p.eager')).toHaveText('eager');
+		expect(page.locator('p.loadingsuccess')).toBeVisible();
+		expect(page.locator('p.loadingfail')).toBeVisible();
+
+		await expect(page.locator('p.success')).toHaveText('success');
+		await expect(page.locator('p.fail')).toHaveText('fail');
+		expect(page.locator('p.loadingsuccess')).toBeHidden();
+		expect(page.locator('p.loadingfail')).toBeHidden();
+	});
+
+	test('Works for server load functions (client nav)', async ({ page }) => {
+		await page.goto('/streaming');
+		page.click('[href="/streaming/server"]');
+
+		await expect(page.locator('p.eager')).toHaveText('eager');
+		expect(page.locator('p.loadingsuccess')).toBeVisible();
+		expect(page.locator('p.loadingfail')).toBeVisible();
+
+		await expect(page.locator('p.success', { timeout: 15000 })).toHaveText('success');
+		await expect(page.locator('p.fail', { timeout: 15000 })).toHaveText('fail');
+		expect(page.locator('p.loadingsuccess')).toBeHidden();
+		expect(page.locator('p.loadingfail')).toBeHidden();
+	});
+
+	// TODO `vite preview` buffers responses, causing these tests to fail
+	if (process.env.DEV) {
+		test('Works for universal load functions (direct hit)', async ({ page }) => {
+			page.goto('/streaming/universal');
+
+			// Write first assertion like this to control the retry interval. Else it might happen that
+			// the test fails because the next retry is too late (probably uses a back-off strategy)
+			await expect(async () => {
+				expect(await page.locator('p.eager').textContent()).toBe('eager');
+			}).toPass({
+				intervals: [100]
+			});
+
+			expect(page.locator('p.loadingsuccess')).toBeVisible();
+			expect(page.locator('p.loadingfail')).toBeVisible();
+
+			await expect(page.locator('p.success')).toHaveText('success');
+			await expect(page.locator('p.fail')).toHaveText('fail');
+			expect(page.locator('p.loadingsuccess')).toBeHidden();
+			expect(page.locator('p.loadingfail')).toBeHidden();
+		});
+
+		test('Works for server load functions (direct hit)', async ({ page }) => {
+			page.goto('/streaming/server');
+
+			// Write first assertion like this to control the retry interval. Else it might happen that
+			// the test fails because the next retry is too late (probably uses a back-off strategy)
+			await expect(async () => {
+				expect(await page.locator('p.eager').textContent()).toBe('eager');
+			}).toPass({
+				intervals: [100]
+			});
+
+			expect(page.locator('p.loadingsuccess')).toBeVisible();
+			expect(page.locator('p.loadingfail')).toBeVisible();
+
+			await expect(page.locator('p.success')).toHaveText('success');
+			await expect(page.locator('p.fail')).toHaveText('fail');
+			expect(page.locator('p.loadingsuccess')).toBeHidden();
+			expect(page.locator('p.loadingfail')).toBeHidden();
+		});
+	}
 });